fix types for Switch node
