+++ src/nodebpy/nodes/geometry/__init__.py
@@ -5,6 +5,7 @@ from .manual import (
     HandleTypeSelection,
     IndexSwitch,
     MenuSwitch,
+    Switch,
     CaptureAttribute,
     FieldToGrid,
     JoinGeometry,
@@ -98,7 +99,6 @@ from .converter import (
     StoreBundleItem,
     StringLength,
     StringToValue,
-    Switch,
     TransformDirection,
     TransformPoint,
     TransposeMatrix,
