+++ src/nodebpy/nodes/geometry/converter.py
@@ -5162,307 +5162,6 @@ class StringToValue(BaseNode):
         self.node.data_type = value
 
 
-class Switch(BaseNode):
-    """
-    Switch between two inputs
-
-    Parameters
-    ----------
-    switch : InputBoolean
-        Switch
-    false : InputFloat
-        False
-    true : InputFloat
-        True
-
-    Inputs
-    ------
-    i.switch : BooleanSocket
-        Switch
-    i.false : FloatSocket
-        False
-    i.true : FloatSocket
-        True
-
-    Outputs
-    -------
-    o.output : FloatSocket
-        Output
-    """
-
-    _bl_idname = "GeometryNodeSwitch"
-    node: bpy.types.GeometryNodeSwitch
-
-    class _Inputs(SocketAccessor):
-        switch: BooleanSocket
-        """Switch"""
-        false: FloatSocket
-        """False"""
-        true: FloatSocket
-        """True"""
-
-    class _Outputs(SocketAccessor):
-        output: FloatSocket
-        """Output"""
-
-    if TYPE_CHECKING:
-
-        @property
-        def i(self) -> _Inputs: ...
-        @property
-        def o(self) -> _Outputs: ...
-
-    def __init__(
-        self,
-        switch: InputBoolean = False,
-        false: InputFloat = 0.0,
-        true: InputFloat = 0.0,
-        *,
-        input_type: Literal[
-            "FLOAT",
-            "INT",
-            "BOOLEAN",
-            "VECTOR",
-            "RGBA",
-            "ROTATION",
-            "MATRIX",
-            "STRING",
-            "MENU",
-            "OBJECT",
-            "IMAGE",
-            "GEOMETRY",
-            "COLLECTION",
-            "MATERIAL",
-            "BUNDLE",
-            "CLOSURE",
-            "FONT",
-        ] = "FLOAT",
-    ):
-        super().__init__()
-        key_args = {"Switch": switch, "False": false, "True": true}
-        self.input_type = input_type
-        self._establish_links(**key_args)
-
-    @classmethod
-    def float(
-        cls,
-        switch: InputBoolean = False,
-        false: InputFloat = 0.0,
-        true: InputFloat = 0.0,
-    ) -> "Switch":
-        """Create Switch with operation 'Float'."""
-        return cls(input_type="FLOAT", switch=switch, false=false, true=true)
-
-    @classmethod
-    def integer(
-        cls,
-        switch: InputBoolean = False,
-        false: InputInteger = 0,
-        true: InputInteger = 0,
-    ) -> "Switch":
-        """Create Switch with operation 'Integer'."""
-        return cls(input_type="INT", switch=switch, false=false, true=true)
-
-    @classmethod
-    def boolean(
-        cls,
-        switch: InputBoolean = False,
-        false: InputBoolean = False,
-        true: InputBoolean = False,
-    ) -> "Switch":
-        """Create Switch with operation 'Boolean'."""
-        return cls(input_type="BOOLEAN", switch=switch, false=false, true=true)
-
-    @classmethod
-    def vector(
-        cls,
-        switch: InputBoolean = False,
-        false: InputVector = None,
-        true: InputVector = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Vector'."""
-        return cls(input_type="VECTOR", switch=switch, false=false, true=true)
-
-    @classmethod
-    def color(
-        cls,
-        switch: InputBoolean = False,
-        false: InputColor = None,
-        true: InputColor = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Color'."""
-        return cls(input_type="RGBA", switch=switch, false=false, true=true)
-
-    @classmethod
-    def rotation(
-        cls,
-        switch: InputBoolean = False,
-        false: InputRotation = None,
-        true: InputRotation = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Rotation'."""
-        return cls(input_type="ROTATION", switch=switch, false=false, true=true)
-
-    @classmethod
-    def matrix(
-        cls,
-        switch: InputBoolean = False,
-        false: InputMatrix = None,
-        true: InputMatrix = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Matrix'."""
-        return cls(input_type="MATRIX", switch=switch, false=false, true=true)
-
-    @classmethod
-    def string(
-        cls,
-        switch: InputBoolean = False,
-        false: InputString = "",
-        true: InputString = "",
-    ) -> "Switch":
-        """Create Switch with operation 'String'."""
-        return cls(input_type="STRING", switch=switch, false=false, true=true)
-
-    @classmethod
-    def menu(
-        cls,
-        switch: InputBoolean = False,
-        false: InputMenu = None,
-        true: InputMenu = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Menu'."""
-        return cls(input_type="MENU", switch=switch, false=false, true=true)
-
-    @classmethod
-    def object(
-        cls,
-        switch: InputBoolean = False,
-        false: InputObject = None,
-        true: InputObject = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Object'."""
-        return cls(input_type="OBJECT", switch=switch, false=false, true=true)
-
-    @classmethod
-    def image(
-        cls,
-        switch: InputBoolean = False,
-        false: InputImage = None,
-        true: InputImage = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Image'."""
-        return cls(input_type="IMAGE", switch=switch, false=false, true=true)
-
-    @classmethod
-    def geometry(
-        cls,
-        switch: InputBoolean = False,
-        false: InputGeometry = None,
-        true: InputGeometry = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Geometry'."""
-        return cls(input_type="GEOMETRY", switch=switch, false=false, true=true)
-
-    @classmethod
-    def collection(
-        cls,
-        switch: InputBoolean = False,
-        false: InputCollection = None,
-        true: InputCollection = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Collection'."""
-        return cls(input_type="COLLECTION", switch=switch, false=false, true=true)
-
-    @classmethod
-    def material(
-        cls,
-        switch: InputBoolean = False,
-        false: InputMaterial = None,
-        true: InputMaterial = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Material'."""
-        return cls(input_type="MATERIAL", switch=switch, false=false, true=true)
-
-    @classmethod
-    def bundle(
-        cls,
-        switch: InputBoolean = False,
-        false: InputBundle = None,
-        true: InputBundle = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Bundle'."""
-        return cls(input_type="BUNDLE", switch=switch, false=false, true=true)
-
-    @classmethod
-    def closure(
-        cls,
-        switch: InputBoolean = False,
-        false: InputClosure = None,
-        true: InputClosure = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Closure'."""
-        return cls(input_type="CLOSURE", switch=switch, false=false, true=true)
-
-    @classmethod
-    def font(
-        cls,
-        switch: InputBoolean = False,
-        false: InputFont = None,
-        true: InputFont = None,
-    ) -> "Switch":
-        """Create Switch with operation 'Font'."""
-        return cls(input_type="FONT", switch=switch, false=false, true=true)
-
-    @property
-    def input_type(
-        self,
-    ) -> Literal[
-        "FLOAT",
-        "INT",
-        "BOOLEAN",
-        "VECTOR",
-        "RGBA",
-        "ROTATION",
-        "MATRIX",
-        "STRING",
-        "MENU",
-        "OBJECT",
-        "IMAGE",
-        "GEOMETRY",
-        "COLLECTION",
-        "MATERIAL",
-        "BUNDLE",
-        "CLOSURE",
-        "FONT",
-    ]:
-        return self.node.input_type
-
-    @input_type.setter
-    def input_type(
-        self,
-        value: Literal[
-            "FLOAT",
-            "INT",
-            "BOOLEAN",
-            "VECTOR",
-            "RGBA",
-            "ROTATION",
-            "MATRIX",
-            "STRING",
-            "MENU",
-            "OBJECT",
-            "IMAGE",
-            "GEOMETRY",
-            "COLLECTION",
-            "MATERIAL",
-            "BUNDLE",
-            "CLOSURE",
-            "FONT",
-        ],
-    ):
-        self.node.input_type = value
-
-
 class TransformDirection(BaseNode):
     """
     Apply a transformation matrix (excluding translation) to the given vector
