+++ src/nodebpy/nodes/geometry/manual.py
@@ -1,15 +1,15 @@
-import bpy.types
 from collections.abc import Mapping
 from typing import TYPE_CHECKING, Any, Generic, Iterable, Literal, TypeVar
 
 import bpy
+import bpy.types
 from bpy.types import (
+    ColorRamp,
     ColorRampElements,
     CurveMapPoints,
     NodeEvaluateClosure,
     NodeSocket,
     NodeSocketString,
-    ColorRamp,
 )
 
 from nodebpy.builder._registry import _get_socket_linker
@@ -23,6 +23,7 @@ from ...builder import (
     ColorSocket,
     DynamicInputsMixin,
     FloatSocket,
+    FontSocket,
     GeometrySocket,
     ImageSocket,
     IntegerSocket,
@@ -48,6 +49,7 @@ from ...types import (
     InputCollection,
     InputColor,
     InputFloat,
+    InputFont,
     InputGeometry,
     InputGrid,
     InputImage,
@@ -69,7 +71,6 @@ from ...types import (
     _GridDataTypes,
     _is_default_value,
 )
-from .converter import Switch
 from .zone import (
     ClosureInput,
     ClosureOutput,
@@ -133,6 +134,7 @@ __all__ = (
     "Float",
     "FloatCurve",
     "ColorRamp",
+    "Switch",
     "StoreNamedAttribute",
 )
 
@@ -144,6 +146,274 @@ def tree(
     arrange: Literal["sugiyama", "simple"] | None = "sugiyama",
 ) -> TreeBuilder:
     return TreeBuilder.geometry(name, collapse=collapse, arrange=arrange)
+
+
+_SwitchDataTypes = Literal[
+    "FLOAT",
+    "INT",
+    "BOOLEAN",
+    "VECTOR",
+    "RGBA",
+    "ROTATION",
+    "MATRIX",
+    "STRING",
+    "MENU",
+    "OBJECT",
+    "IMAGE",
+    "GEOMETRY",
+    "COLLECTION",
+    "MATERIAL",
+    "BUNDLE",
+    "CLOSURE",
+    "FONT",
+]
+
+
+class Switch(BaseNode, Generic[_T]):
+    """
+    Switch between two inputs
+
+    Parameters
+    ----------
+    switch : InputBoolean
+        Switch
+    false : InputFloat
+        False
+    true : InputFloat
+        True
+
+    Inputs
+    ------
+    i.switch : BooleanSocket
+        Switch
+    i.false : FloatSocket
+        False
+    i.true : FloatSocket
+        True
+
+    Outputs
+    -------
+    o.output : FloatSocket
+        Output
+    """
+
+    _bl_idname = "GeometryNodeSwitch"
+    node: bpy.types.GeometryNodeSwitch
+
+    class _Inputs(SocketAccessor, Generic[_S]):
+        switch: BooleanSocket
+        """Switch"""
+        false: _S
+        """False"""
+        true: _S
+        """True"""
+
+    class _Outputs(SocketAccessor, Generic[_S]):
+        output: _S
+        """Output"""
+
+    if TYPE_CHECKING:
+
+        @property
+        def i(self) -> _Inputs: ...
+        @property
+        def o(self) -> _Outputs: ...
+
+    def __init__(
+        self,
+        switch: InputBoolean = False,
+        false: _S = 0.0,
+        true: _S = 0.0,
+        *,
+        input_type: _SwitchDataTypes = "FLOAT",
+    ):
+        super().__init__()
+        key_args = {"Switch": switch, "False": false, "True": true}
+        self.input_type = input_type
+        self._establish_links(**key_args)
+
+    @classmethod
+    def float(
+        cls,
+        switch: InputBoolean = False,
+        false: InputFloat = 0.0,
+        true: InputFloat = 0.0,
+    ) -> "Switch[FloatSocket]":
+        """Create Switch with operation 'Float'."""
+        return Switch(input_type="FLOAT", switch=switch, false=false, true=true)
+
+    @classmethod
+    def integer(
+        cls,
+        switch: InputBoolean = False,
+        false: InputInteger = 0,
+        true: InputInteger = 0,
+    ) -> "Switch[IntegerSocket]":
+        """Create Switch with operation 'Integer'."""
+        return Switch(input_type="INT", switch=switch, false=false, true=true)
+
+    @classmethod
+    def boolean(
+        cls,
+        switch: InputBoolean = False,
+        false: InputBoolean = False,
+        true: InputBoolean = False,
+    ) -> "Switch[BooleanSocket]":
+        """Create Switch with operation 'Boolean'."""
+        return Switch(input_type="BOOLEAN", switch=switch, false=false, true=true)
+
+    @classmethod
+    def vector(
+        cls,
+        switch: InputBoolean = False,
+        false: InputVector = None,
+        true: InputVector = None,
+    ) -> "Switch[VectorSocket]":
+        """Create Switch with operation 'Vector'."""
+        return Switch(input_type="VECTOR", switch=switch, false=false, true=true)
+
+    @classmethod
+    def color(
+        cls,
+        switch: InputBoolean = False,
+        false: InputColor = None,
+        true: InputColor = None,
+    ) -> "Switch[ColorSocket]":
+        """Create Switch with operation 'Color'."""
+        return Switch(input_type="RGBA", switch=switch, false=false, true=true)
+
+    @classmethod
+    def rotation(
+        cls,
+        switch: InputBoolean = False,
+        false: InputRotation = None,
+        true: InputRotation = None,
+    ) -> "Switch[RotationSocket]":
+        """Create Switch with operation 'Rotation'."""
+        return Switch(input_type="ROTATION", switch=switch, false=false, true=true)
+
+    @classmethod
+    def matrix(
+        cls,
+        switch: InputBoolean = False,
+        false: InputMatrix = None,
+        true: InputMatrix = None,
+    ) -> "Switch[MatrixSocket]":
+        """Create Switch with operation 'Matrix'."""
+        return Switch(input_type="MATRIX", switch=switch, false=false, true=true)
+
+    @classmethod
+    def string(
+        cls,
+        switch: InputBoolean = False,
+        false: InputString = "",
+        true: InputString = "",
+    ) -> "Switch[StringSocket]":
+        """Create Switch with operation 'String'."""
+        return Switch(input_type="STRING", switch=switch, false=false, true=true)
+
+    @classmethod
+    def menu(
+        cls,
+        switch: InputBoolean = False,
+        false: InputMenu = None,
+        true: InputMenu = None,
+    ) -> "Switch[MenuSocket]":
+        """Create Switch with operation 'Menu'."""
+        return Switch(input_type="MENU", switch=switch, false=false, true=true)
+
+    @classmethod
+    def object(
+        cls,
+        switch: InputBoolean = False,
+        false: InputObject = None,
+        true: InputObject = None,
+    ) -> "Switch[ObjectSocket]":
+        """Create Switch with operation 'Object'."""
+        return Switch(input_type="OBJECT", switch=switch, false=false, true=true)
+
+    @classmethod
+    def image(
+        cls,
+        switch: InputBoolean = False,
+        false: InputImage = None,
+        true: InputImage = None,
+    ) -> "Switch[ImageSocket]":
+        """Create Switch with operation 'Image'."""
+        return Switch(input_type="IMAGE", switch=switch, false=false, true=true)
+
+    @classmethod
+    def geometry(
+        cls,
+        switch: InputBoolean = False,
+        false: InputGeometry = None,
+        true: InputGeometry = None,
+    ) -> "Switch[GeometrySocket]":
+        """Create Switch with operation 'Geometry'."""
+        return Switch(input_type="GEOMETRY", switch=switch, false=false, true=true)
+
+    @classmethod
+    def collection(
+        cls,
+        switch: InputBoolean = False,
+        false: InputCollection = None,
+        true: InputCollection = None,
+    ) -> "Switch[CollectionSocket]":
+        """Create Switch with operation 'Collection'."""
+        return Switch(input_type="COLLECTION", switch=switch, false=false, true=true)
+
+    @classmethod
+    def material(
+        cls,
+        switch: InputBoolean = False,
+        false: InputMaterial = None,
+        true: InputMaterial = None,
+    ) -> "Switch[MaterialSocket]":
+        """Create Switch with operation 'Material'."""
+        return Switch(input_type="MATERIAL", switch=switch, false=false, true=true)
+
+    @classmethod
+    def bundle(
+        cls,
+        switch: InputBoolean = False,
+        false: InputBundle = None,
+        true: InputBundle = None,
+    ) -> "Switch[BundleSocket]":
+        """Create Switch with operation 'Bundle'."""
+        return Switch(input_type="BUNDLE", switch=switch, false=false, true=true)
+
+    @classmethod
+    def closure(
+        cls,
+        switch: InputBoolean = False,
+        false: InputClosure = None,
+        true: InputClosure = None,
+    ) -> "Switch[ClosureSocket]":
+        """Create Switch with operation 'Closure'."""
+        return Switch(input_type="CLOSURE", switch=switch, false=false, true=true)
+
+    @classmethod
+    def font(
+        cls,
+        switch: InputBoolean = False,
+        false: InputFont = None,
+        true: InputFont = None,
+    ) -> "Switch[FontSocket]":
+        """Create Switch with operation 'Font'."""
+        return Switch(input_type="FONT", switch=switch, false=false, true=true)
+
+    @property
+    def input_type(
+        self,
+    ) -> _SwitchDataTypes:
+        return self.node.input_type  # ty: ignore[invalid-return-type]
+
+    @input_type.setter
+    def input_type(
+        self,
+        value: _SwitchDataTypes,
+    ):
+        self.node.input_type = value
 
 
 _ColorRampColorInterpolations = Literal[
